Add explicit type cast in generate_random()
Without this explicit type cast the function fails on some
setups.

The exact reason is unclear and might depend on obscure compiler
setups, platform defaults etc.

+++ util.c
@@ -207,7 +207,7 @@ int generate_random(void *buf, int len)
 		return -1;
 	}
 
-	res = fread(buf, 1, (size_t) len, u);
+	res = (int) fread(buf, 1, (size_t) len, u);
 	fclose(u);
 
 	return (res != len);
